# Conflicts:
#	src/Model/Battle.java

+++ src/Controller/Controller.java
@@ -37,7 +37,7 @@ public class Controller {
                 save();
                 break;
             case LOGOUT:
-                logout(request);
+                logout();
                 break;
             case HELP:
                 help();
@@ -81,11 +81,8 @@ public class Controller {
             case SHOW_DECK:
                 showDeck(request);
                 break;
-            case SHOW_COLLECTION_IN_SHOP:
-                showCollectionShop(request);
-                break;
             case SEARCH:
-                searchInShop(request);
+                search(request);
                 break;
             case BUY:
                 buy(request);
@@ -109,7 +106,7 @@ public class Controller {
                 showCardInfo(request);
                 break;
             case SELECTION:
-                selectCardInBattle(request);
+                select(request);
                 break;
             case MOVE:
                 moveToInBattle(request);
@@ -129,11 +126,35 @@ public class Controller {
             case INSERTION:
                 insertCard(request);
                 break;
+            case END_TURN:
+                endTurn();
+                break;
+            case SHOW_COLLECTABLES:
+                showCollectables();
+                break;
+            case SHOW_COLLECTABLE_INFO:
+                showCollectableInfo();
+                break;
+            case USE_ITEM:
+                useItem(request);
+                break;
+            case NEXT_CARD:
+                showNextCard();
+                break;
+            case SHOW_CARDS:
+                showCards();
+                break;
+            case END_GAME:
+                endGame();
+                break;
+            case SHOW_MENU:
+                showMenu();
+                break;
         }
     }
 
     public void createAccount(Request request) {
-        if (request.checkAccountCreationSyntax()) {
+        if (request.checkAccountCreationSyntax() && menu.getStat() == MenuStat.MAIN) {
             String username = request.getCommand();
             view.passwordInsertion();
             String password = request.getNewCommand();
@@ -143,7 +164,7 @@ public class Controller {
     }
 
     public void login(Request request) {
-        if (request.checkLoginSyntax()) {
+        if (request.checkLoginSyntax() && menu.getStat() == MenuStat.MAIN) {
             String username = request.getCommand();
             view.passwordInsertion();
             String password = request.getNewCommand();
@@ -152,17 +173,19 @@ public class Controller {
     }
 
     public void showLeaderBoard(Request request) {
-        if (request.checkLeaderBoardSyntax()) {
+        if (request.checkLeaderBoardSyntax() && menu.getStat() == MenuStat.ACCOUNT) {
             view.printLeaderboard();
         }
     }
 
     public void save() {
+        if (menu.getStat() == MenuStat.ACCOUNT) {
 
+        }
     }
 
-    public void logout(Request request) {
-        if (request.checkLogoutSyntax()) {
+    public void logout() {
+        if (menu.getStat() == MenuStat.ACCOUNT) {
             view.logout();
         }
     }
@@ -191,16 +214,19 @@ public class Controller {
     }
 
     public void showTheCollection() {
-        view.printCollection(this.account.getCollection());
+        if (menu.getStat() == MenuStat.COLLECTION && menu.getStat() == MenuStat.SHOP)
+            view.printCollection(this.account.getCollection());
     }
 
 
     public void searchInCollection(Request request) {
         if (request.checkSearchSyntax()) {
-            view.printId(Card.getAllCardsId(request.getObjectName(request.getCommand()),
-                    this.account.getCollection().getCards().toArray(Card[]::new)).toArray(Card[]::new));
-            view.printId(Item.getAllItemsId(request.getObjectName(request.getCommand()),
-                    this.account.getCollection().getItems().toArray(Item[]::new)).toArray(Item[]::new));
+            if (menu.getStat() == MenuStat.COLLECTION || menu.getStat() == MenuStat.SHOP) {
+                view.printId(Card.getAllCardsId(request.getObjectName(request.getCommand()),
+                        this.account.getCollection().getCards().toArray(Card[]::new)).toArray(Card[]::new));
+                view.printId(Item.getAllItemsId(request.getObjectName(request.getCommand()),
+                        this.account.getCollection().getItems().toArray(Item[]::new)).toArray(Item[]::new));
+            }
         }
     }
 
@@ -209,101 +235,109 @@ public class Controller {
     }
 
     public void createDeck(Request request) {
-        if (request.checkDeckSyntax()) {
+        if (request.checkDeckSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.createDeck(this.account.getCollection().createDeck(request.getDeckName(request.getCommand())));
         }
     }
 
     public void deleteDeck(Request request) {
-        if (request.checkDeckSyntax()) {
+        if (request.checkDeckSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.deleteDeck(this.account.getCollection().deleteDeck(request.getDeckName(request.getCommand())));
         }
     }
 
     public void addToDeck(Request request) {
-        if (request.checkToDeckAdditionSyntax()) {
+        if (request.checkToDeckAdditionSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.addToCollection(this.account.getCollection().add(request.getDeckName(request.getCommand()), request.getObjectID(request.getCommand())));
         }
     }
 
     public void removeFromDeck(Request request) {
-        if (request.checkFromDeckDeletionSyntax()) {
+        if (request.checkFromDeckDeletionSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.removeFromDeck(this.account.getCollection().remove(request.getDeckName(request.getCommand()), request.getObjectID(request.getCommand())));
         }
     }
 
     public void validateDeck(Request request) {
-        if (request.checkValidationSyntax()) {
+        if (request.checkValidationSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.checkValidation(this.account.getCollection().validate(request.getDeckName(request.getCommand())));
         }
     }
 
     public void selectDeck(Request request) {
-        if (request.checkDeckSelectionSyntax()) {
+        if (request.checkDeckSelectionSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.printDeckSelection(this.account.getCollection().selectDeck(request.getDeckName(request.getCommand())));
         }
     }
 
     public void showAllDecks(Request request) {
-        if (request.checkShowAllDeckSyntax()) {
+        if (request.checkShowAllDeckSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.showAllDeck(this.account.getCollection().getDecks());
         }
     }
 
     public void showDeck(Request request) {
-        if (request.checkShowDeckSyntax()) {
+        if (request.checkShowDeckSyntax() && menu.getStat() == MenuStat.COLLECTION) {
             view.printDeck(this.account.getCollection().getDecks().get(
                     this.account.getCollection().deckExistance(request.getDeckName(request.getCommand()))));
         }
     }
 
-    public void showCollectionShop(Request request) {
-        view.printShopCollection(this.account.getCollection());
-    }
-
-    public void searchInShop(Request request) {
+    public void search(Request request) {
         if (request.checkSearchSyntax()) {
-            Card card = Card.getCardByName(request.getObjectName(request.getCommand()),
-                    shop.getCards().toArray(Card[]::new));
-            Item item = Item.getItemByName(request.getObjectName(request.getCommand()),
-                    shop.getItems().toArray(Item[]::new));
-            if (card != null)
-                view.printId(card);
-            if (item != null)
-                view.printId(item);
+            if (menu.getStat() == MenuStat.COLLECTION) {
+                searchInCollection(request);
+            } else if (menu.getStat() == MenuStat.SHOP) {
+                Card card = Card.getCardByName(request.getObjectName(request.getCommand()),
+                        shop.getCards().toArray(Card[]::new));
+                Item item = Item.getItemByName(request.getObjectName(request.getCommand()),
+                        shop.getItems().toArray(Item[]::new));
+                if (card != null)
+                    view.printId(card);
+                if (item != null)
+                    view.printId(item);
+            }
         }
     }
 
     public void buy(Request request) {
-        if (request.checkBuySyntax()) {
+        if (request.checkBuySyntax() && menu.getStat() == MenuStat.SHOP) {
             view.printBuyMessages(shop.buy(request.getObjectName(request.getCommand()), this.account));
         }
     }
 
     public void sell(Request request) {
-        if (request.checkSellSyntax()) {
+        if (request.checkSellSyntax() && menu.getStat() == MenuStat.SHOP) {
             view.printSellMessages(shop.sell(request.getObjectID(request.getCommand()), this.account));
         }
     }
 
     public void showShop() {
-        view.printShopCollection(new Collection(shop.getCards(), shop.getItems()));
+        if (menu.getStat() == MenuStat.SHOP) {
+            view.printShopCollection(new Collection(shop.getCards(), shop.getItems()));
+        }
     }
 
     public void gameInfo() {
+        if (menu.getStat() == MenuStat.BATTLE) {
 
+        }
     }
 
     public void showMyMinions() {
-        view.printMinionsInfo(battle.getFieldCards()[battle.getTurnByAccount(this.account)]);
+        if (menu.getStat() == MenuStat.BATTLE) {
+            view.printMinionsInfo(battle.getFieldCards()[battle.getTurnByAccount(this.account)]);
+        }
     }
 
     public void showOppMinions() {
-        view.printMinionsInfo(battle.getFieldCards()[(battle.getTurnByAccount(this.account) + 1) % 2]);
+        if (menu.getStat() == MenuStat.BATTLE) {
+            view.printMinionsInfo(battle.getFieldCards()[(battle.getTurnByAccount(this.account) + 1) % 2]);
+        }
     }
 
     public void showCardInfo(Request request) {
-        if (request.checkFetchInfoSyntax()) {
+        if (request.checkFetchInfoSyntax() && menu.getStat() == MenuStat.BATTLE) {
             Card card = Card.getCardByID(request.getObjectID(request.getCommand()),
                     battle.getPlayerHands()[battle.getTurnByAccount(account)]);
             if (card != null) {
@@ -316,23 +350,29 @@ public class Controller {
         }
     }
 
-    public void selectCardInBattle(Request request) {
-        if (request.checkCardSelectionSyntax()) {
-            Card card = Card.getCardByID(request.getObjectID(request.getCommand()),
-                    battle.getFieldCards()[battle.getTurnByAccount(account)]);
-            view.printCardInfo(card);
+    public void select(Request request) {
+        if (request.checkCardSelectionSyntax() && menu.getStat() == MenuStat.BATTLE) {
+            int turn = battle.getTurnByAccount(account);
+            Card card = Card.getCardByID(request.getObjectID(request.getCommand()), battle.getFieldCards()[turn]);
+            Item item = Item.getItemByID(request.getObjectID(request.getCommand()), battle.getCollectables()[turn]);
+            if (card != null)
+                battle.selectCard(card.getId());
+            else if (item != null)
+                battle.selectItem(item.getId());
+            else
+                view.printUnsuccessfulSelection(battle.getCollectables()[turn].length);
         }
     }
 
     public void moveToInBattle(Request request) {
-        if (request.checkMoveSyntax()) {
+        if (request.checkMoveSyntax() && menu.getStat() == MenuStat.BATTLE) {
             Coordinate coordinate = request.getCoordinate(request.getCommand());
             view.showMovement(battle.moveTo(coordinate));
         }
     }
 
     public void battleAttack(Request request) {
-        if (request.checkAssaultSyntax()) {
+        if (request.checkAssaultSyntax() && menu.getStat() == MenuStat.BATTLE) {
             Card card = Card.getCardByID(request.getObjectID(request.getCommand()),
                     battle.getFieldCards()[(battle.getTurnByAccount(account) + 1) % 2]);
             view.showAttack(battle.attack(card.getId(),
@@ -341,7 +381,7 @@ public class Controller {
     }
 
     public void battleComboAttack(Request request) {
-        if (request.checkComboSyntax()) {
+        if (request.checkComboSyntax() && menu.getStat() == MenuStat.BATTLE) {
             int oppId = request.getOppIdInCombo(request.getCommand());
             int[] ids = request.getComboComradesId(request.getCommand());
             Card[] cards = new Card[ids.length];
@@ -353,60 +393,71 @@ public class Controller {
     }
 
     public void useSpecialPower(Request request) {
-        if (request.checkSPUsageSyntax()) {
+        if (request.checkSPUsageSyntax() && menu.getStat() == MenuStat.BATTLE) {
             Coordinate target = request.getCoordinate(request.getCommand());
 
         }
     }
 
     public void showHand() {
-        view.printHand(battle.getPlayerHands()[battle.getTurnByAccount(account)]);
+        if (menu.getStat() == MenuStat.BATTLE) {
+            view.printHand(battle.getPlayerHands()[battle.getTurnByAccount(account)]);
+        }
     }
 
     public void insertCard(Request request) {
-        if (request.checkCardInsertSyntax()) {
+        if (request.checkCardInsertSyntax() && menu.getStat() == MenuStat.BATTLE) {
             view.printInsertionMessage(battle.insertCard(request.getCoordinate(request.getCommand()),
                     request.getInsertedName(request.getCommand())));
         }
     }
 
     public void endTurn() {
-
+        if (menu.getStat() == MenuStat.BATTLE) {
+            battle.endTurn();
+            view.endTurn();
+        }
     }
 
     public void showCollectables() {
-
+        if (menu.getStat() == MenuStat.BATTLE) {
+            view.printCollectables(battle.getCollectables()[battle.getTurnByAccount(account)]);
+        }
     }
 
-    public void selectCollectables() {
-
+    private void showCollectableInfo() {
+        if (menu.getStat() == MenuStat.ITEM_SELECTION) {
+            view.printItem(battle.getCurrentItem());
+        }
     }
 
-    public void showBattleInfo() {
-
-    }
-
-    public void useItem() {
-
+    public void useItem(Request request) {
+        if (request.checkItemUseSyntax()) {
+            if (menu.getStat() == MenuStat.BATTLE || menu.getStat() == MenuStat.ITEM_SELECTION) {
+                view.printItemUsage(battle.useItem);
+            }
+        }
     }
 
     public void showNextCard() {
-
-    }
-
-    public void enterTheGraveyard() {
-
+        if (menu.getStat() == MenuStat.BATTLE) {
+            view.printCardInfo(account.getCollection().getMainDeck().getCards().get(0));
+        }
     }
 
     public void showCards() {
-
+        if (menu.getStat() == MenuStat.GRAVEYARD) {
+            view.printCards(battle.getGraveyard()[battle.getTurnByAccount(account)]);
+        }
     }
 
     public void endGame() {
-
+        if (menu.getStat() == MenuStat.BATTLE) {
+            view.endGame(battle);
+        }
     }
 
     public void showMenu() {
-
+        view.printOptions();
     }
 }
